rename components and fix deposit input tests
