+++ packages/ui-components/src/lib/index.ts
@@ -46,8 +46,8 @@ export { default as IconSuccess } from './components/IconSuccess.svelte';
 export { default as IconTelegram } from './components/IconTelegram.svelte';
 export { default as IconWalletConnect } from './components/IconWalletConnect.svelte';
 export { default as IconWarning } from './components/IconWarning.svelte';
-export { default as FieldDefinitionButtons } from './components/deployment/FieldDefinitionButtons.svelte';
-export { default as DepositButtons } from './components/deployment/DepositButtons.svelte';
+export { default as FieldDefinitionInput } from './components/deployment/FieldDefinitionInput.svelte';
+export { default as DepositInput } from './components/deployment/DepositInput.svelte';
 export { default as DeploymentSteps } from './components/deployment/DeploymentSteps.svelte';
 export { default as TokenOutput } from './components/deployment/TokenInputOrOutput.svelte';
 export { default as SelectToken } from './components/deployment/SelectToken.svelte';
